モデルフィールドの修正

+++ map_info/map_app/models.py
@@ -1,8 +1,8 @@
 from django.db import models
 
 class Positions(models.Model):
-    lat = models.IntegerField()
-    lng = models.IntegerField()
+    lat = models.DecimalField(max_digits=10, decimal_places=7)
+    lng = models.DecimalField(max_digits=10, decimal_places=7)
 
 class ParkInfo(models.Model):
     park_name = models.CharField(max_length=50, default='公園')
